requirements.txt
Babel==2.2.0
beautifulsoup4==4.4.1
csscompressor==0.9.4
Django==1.9.2
django-appconf==1.0.1
django-compressor==2.0
django-debug-toolbar==1.4
django-modelcluster==1.1
django-taggit==0.18.0
django-treebeard==3.0
djangorestframework==3.3.2
easy-thumbnails==2.3
futures==3.0.4
gunicorn==19.4.5
html5lib==0.999
Jinja2==2.8
lxml==3.3.3
MarkupSafe==0.23
Pillow==3.1.1
ply==3.8
psycopg2==2.6.1
pytz==2015.7
rcssmin==1.0.6
rjsmin==1.0.12
simplejson==3.8.1
six==1.10.0
slimit==0.8.1
sqlparse==0.1.18
Unidecode==0.4.19
wagtail==1.3.1
wheel==0.29.0
Willow==0.2.2
